Value for internationalized target attributes must be unordered
In order to addresshttps://github.com/ietf-wg-httpapi/linkset/issues/2 : Changed the language to remove the suggestion that the JSON objects used in the array for inernationalized target attributes required ordering of members. A JSON object is defined as unordered collection of zero or more name/value pairs and as such suggesting ordering is inappropriate.