Addressing #45 and #47
Reworked Appendix A with a new example that does not leverage IANA-registered link relation types as per the discussion in #45. Also adding the JSON-LD 1.1 solution to the problem of not being able to use "linkset" links in link sets when mapping to JSON-LD as provided by @azaroth42 in #47. Unfortunately, as things stand with not being able to express IANA-registered link relation types as HTTP URIs (see also #45) it will not be possible to use "linkset" links (or any other IANA-registered link types) in link sets that need to be mapped to JSON-LD until a solution is found with that regard.