Addressed title language issue
Addressed an issue regarding the language of the "title" that was raised  by Benjamin Haduk as part of the IESG Evaluation. Made clear that title language is not inherited from Content-Language (and why) and recommended using "title*" instead of "title" because "title*" allows to explicitly provide title language as part of the attribute value. In order not to have to repeat this (and  another aspect related to use of link sets outside the context of an HTTP interaction) in both sections 4.1. and 4.2, the info is provided in the opening of section 4. This required moving some existing language from 4.1 and 4.2 to the intro of 4.