Addressed feedback from Paul Wauters
Addressed feedback received from Paul Wauters as part of the IESG Evaluation. Made the language in the IANA Considerations section consistent for all registration requests. Don't know how to address the URL rendering issues. URLs are consistent/correct in the source document so this is likely an artifact of the rendering engine.